fix(ui): replace query component with hook
