+++ ui/src/app/shared/components/query.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
-import {BehaviorSubject, type Observable} from 'rxjs';
-import {map} from 'rxjs/operators';
+import {type Observable} from 'rxjs';
 
-import {Context} from '../context';
+import {useQuery, useObservableQuery} from '../hooks/query';
 
 export interface QueryProps {
     children: (params: URLSearchParams) => React.ReactNode;
@@ -10,35 +9,6 @@ export interface QueryProps {
 
 export interface ObservableQueryProps {
     children: (params: Observable<URLSearchParams>) => React.ReactNode;
-}
-
-const useQuery = () => {
-    const context = React.useContext(Context);
-
-    return new URLSearchParams(context?.history?.location.search);
-};
-
-function useObservableQuery() {
-    const context = React.useContext(Context);
-
-    const search$ = React.useMemo(() => new BehaviorSubject(context.history.location.search), []);
-
-    const searchParams$ = React.useMemo(() => search$.pipe(map(searchStr => new URLSearchParams(searchStr))), [search$]);
-
-    React.useEffect(() => {
-        const unsubscribe = context.history.listen(location => {
-            search$.next(location.search);
-        });
-        return unsubscribe;
-    }, [context.history, search$]);
-
-    React.useEffect(() => {
-        return () => {
-            search$.complete();
-        };
-    }, [search$]);
-
-    return searchParams$;
 }
 
 function Query({children}: QueryProps) {
@@ -53,4 +23,4 @@ function ObservableQuery({children}: ObservableQueryProps) {
     return <>{children(searchParams$)}</>;
 }
 
-export {ObservableQuery, Query, useQuery, useObservableQuery};
+export {ObservableQuery, Query};
